+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/), adapted for int
 
 ## [Unreleased]
 
+## [401] - 2026-03-12
+
+- [TASK-584] Fix tusk merge git stash failure when modified files exist outside feature branch scope
+
 ## [400] - 2026-03-12
 
 - [TASK-583] Update dashboard header to use project name